+++ Server/controllers/reviewController.js
@@ -7,6 +7,12 @@ const Comment = db.comments
 
 const addReview= async (req, res) => {
     //let id = req.params.id
+    if (req.user.role != "User" && req.user.role != "Admin") {
+        res.status(403).send({
+            message: "Page access is restricted."
+        });
+        return;
+    }
     const PathExists = await Comment.count({where: {postId: req.params.id1, id: req.params.id2}}) > 0;
     if(!PathExists) {
         res.status(404).send({
@@ -16,7 +22,8 @@ const addReview= async (req, res) => {
     }
     let info = {
         rating: req.body.rating,
-        commentId: req.params.id2
+        commentId: req.params.id2,
+        userId: req.user.id
     }
 
     Review.create(info)
@@ -76,6 +83,12 @@ const getReview = async (req, res) => {
 }
 
 const updateReview = async (req, res) => {
+    if (req.user.role != "User" && req.user.role != "Admin") {
+        res.status(403).send({
+            message: "Page access is restricted."
+        });
+        return;
+    }
     let id3 = req.params.id3
     const PathExists = await Comment.count({where: {postId: req.params.id1, id: req.params.id2}}) > 0;
     const PathExists2 = await Review.count({where: {commentId: req.params.id2, id: req.params.id3}}) > 0;
@@ -85,7 +98,7 @@ const updateReview = async (req, res) => {
         });
         return;
     }
-    Review.update(req.body, {where: {commentId: req.params.id2, id: id3}})
+    Review.update(req.body, {where: {commentId: req.params.id2, id: id3,userId:req.user.id}})
     .then(data => {
         res.status(200).send(data)})
     .catch(err => {
@@ -96,6 +109,12 @@ const updateReview = async (req, res) => {
 }
 
 const deleteReview = async (req, res) => {
+    if (req.user.role != "User" && req.user.role != "Admin") {
+        res.status(403).send({
+            message: "Page access is restricted."
+        });
+        return;
+    }
     const PathExists = await Comment.count({where: {postId: req.params.id1, id: req.params.id2}}) > 0;
     const PathExists2 = await Review.count({where: {commentId: req.params.id2, id: req.params.id3}}) > 0;
     if(!PathExists || !PathExists2) {
@@ -104,13 +123,24 @@ const deleteReview = async (req, res) => {
         });
         return;
     }
-    Review.destroy({where: {id: req.params.id3}})
-    .then(res.status(204).send('review deleted'))
-    .catch(err => {
+    if (req.user.role == "User") {
+        Review.destroy({where: {id: req.params.id3,userId:req.user.id}})
+        .then(res.status(204).send('review deleted'))
+        .catch(err => {
         res.status(500).send({
             message: err.message && "Some error occurred while deleting review."
         });
-    });
+        });
+    }
+    if (req.user.role == "Admin") {
+        Review.destroy({where: {id: req.params.id3}})
+        .then(res.status(204).send('review deleted'))
+        .catch(err => {
+        res.status(500).send({
+            message: err.message && "Some error occurred while deleting review."
+        });
+        });
+    }
 }
 
 module.exports = {
